requirements.txt
fastapi
uvicorn
pyrebase4
pydantic
